% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county}
\alias{county}
\title{VHA geography - county level}
\format{A simple feature collection with 3225 rows and 12 variables:
\describe{
  \item{NAME}{County Name}
  \item{FIPS}{County FIPS Code}
  \item{STATE_NAME}{State where County is located}
  \item{VISN}{Veteran Integrated Service Network ID}
  \item{MARKET}{VHA Market ID}
  \item{SUBMARKET}{VHA Submarket ID}
  \item{SECTOR}{Sector ID}
  \item{STATE_FIPS}{State FIPS Code}
  \item{ST}{State abbreviation}
  \item{FIPS}{County FIPS Code}
  \item{Shape_Leng}{Length of shape}
  \item{Shape_Area}{Area of shape}
  \item{geometry}{Geometry of shape}
}}
\source{
\url{https://www.data.va.gov/dataset/veterans-integrated-services-networks-visn-markets-submarkets-sectors-and-counties}
}
\usage{
county
}
\description{
A dataset of geometry and attributes of VHA geographies at the county level
}
\keyword{datasets}
