% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest_get}
\alias{harvest_get}
\title{set of functions to fetch VHA harvest URLs and dowload in tidy format}
\usage{
harvest_get(...)
}
\arguments{
\item{...}{keyword arguments to make homebrewed API call
possible arguments: VISN, Site, Year, Quarter}
}
\value{
A tibble with download URLs that can be passed on to harvest_wrngl().
        If no arguments are entered, it will return URLs for all available datasets
}
\description{
harvest_get() is a wrapper for dplyr's filter() and queries should be
structured as such.
}
\examples{
\dontrun{
links <- harvest_get(Year == "2018" & VISN == "VISN 2")
links2 <- harves_get(Site \%in\% c("Syracuse", "Buffalo", "Rochester") & Year == 2017)
}

}
