% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest_wrngl}
\alias{harvest_wrngl}
\title{harvest_wrngl() will import and tidy data on VA SAIL metrics.  This function requires
an object created with harvest_get()}
\usage{
harvest_wrngl(data)
}
\arguments{
\item{data}{a data object created with harvest_get()}
}
\value{
A tibble with SAIL metrics in tidy format
}
\description{
harvest_wrngl() will import and tidy data on VA SAIL metrics.  This function requires
an object created with harvest_get()
}
\examples{
\dontrun{
links <- harvest_get(Year == "2018" & VISN == "VISN 2")
data <- harvest_wrngl(links)

This can also be used in a pipe
data <- links \%>\% harvest_wrgnl()
}

}
