% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market}
\alias{market}
\title{VHA geography - market level}
\format{A simple feature collection with 97 rows and 5 variables:
\describe{
  \item{VISN}{Veteran Integrated Service Network ID}
  \item{MARKET}{VHA Market ID}
  \item{Shape_Leng}{Length of shape}
  \item{Shape_Area}{Area of shape}
  \item{geometry}{Geometry of shape}
}}
\usage{
market
}
\description{
A dataset of geometry and attributes of VHA geographies at the market level
}
\keyword{datasets}
