% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sail.R
\name{sail_tidy}
\alias{sail_tidy}
\title{Simple function to tidy Strategic Analytics for Improvement and Learning (SAIL)
datasets from www.va.gov}
\usage{
sail_tidy(url)
}
\arguments{
\item{url}{a SAIL dataset url from: \url{https://www.va.gov/QUALITYOFCARE/measure-up/Strategic_Analytics_for_Improvement_and_Learning_SAIL.asp}}
}
\value{
A tibble with SAIL metrics in tidy format
}
\description{
Simple function to tidy Strategic Analytics for Improvement and Learning (SAIL)
datasets from www.va.gov
}
\examples{
\dontrun{
url <- "https://www.va.gov/QUALITYOFCARE/SAIL_FY19_Q1/SAIL-2_albany_fy19q1.xlsx"
sail_tidy(url)

sail_tidy("https://www.va.gov/QUALITYOFCARE/SAIL_FY19_Q1/SAIL-5_clarksburg_fy19q1.xlsx")
}
}
