% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_geo.R
\name{shift_geo}
\alias{shift_geo}
\title{Shift VHA geographies for Alaska and Hawaii}
\usage{
shift_geo(data)
}
\arguments{
\item{data}{The shapefile with the feature geography you want to shift.
Currently only supports sector and submarket shapefiles.}
}
\value{
An sf tibble of VHA geographies and attributes
}
\description{
Shift VHA geographies for Alaska and Hawaii
}
\examples{
\dontrun{
 shift_geo(submarket) %>%
 ggplot() +
 geom_sf() +
 theme_minimal()
 }

}
\references{
\url{ https://github.com/rdinter/rd3albers}
             \url{https://geocompr.robinlovelace.net/geometric-operations.html}
}
