% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{States}
\format{A simple feature collection with 56 rows and 10 variables:
\describe{
  \item{STATEFP}{State FIPS Code}
  \item{STATENS}{State ANSI Code}
  \item{AFFGEOID}{State American FactFinder GEOID}
  \item{GEOID}{State Geographic Identifier}
  \item{STUSPS}{State Abbreviation}
  \item{NAME}{State Name}
  \item{LSAD}{LSAD Code}
  \item{ALAND}{Land Area}
  \item{AWATER}{Water Area}
  \item{geometry}{Geometry of shape}
}}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html}
}
\usage{
states
}
\description{
A dataset of geometry and attributes of States and Equivalent Entities of the US.
In addition to the 50 states, the District of Columbia, Puerto Rico, American Samoa,
the Commonwealth of the Northern Mariana Islands, Guam, and the U.S. Virgin Islands
are included
}
\keyword{datasets}
