% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Minimal ggplot theme for plotting sf without axis text}
\usage{
theme_map(...)
}
\arguments{
\item{...}{defaults are preset}
}
\description{
Minimal ggplot theme for plotting sf without axis text
}
\examples{
 ggplot() +
 geom_sf(data) +
 theme_map()

}
