% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_va.R
\name{theme_va}
\alias{theme_va}
\title{A precise & pristine [ggplot2] theme with opinionated defaults and an emphasis on typography}
\usage{
theme_va(base_family = "", base_size = 11.5,
  plot_title_family = base_family, plot_title_size = 14,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = base_family, subtitle_size = 12,
  subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", caption_family = base_family,
  caption_size = 9, caption_face = "bold.italic",
  caption_margin = 10, axis_text_size = 9,
  axis_title_family = subtitle_family, axis_title_size = 10,
  axis_title_face = "plain", axis_title_just = "rt",
  plot_margin = margin(15, 15, 15, 15), grid_col = "#cccccc",
  grid = TRUE, map = FALSE, axis_col = "#cccccc", axis = FALSE,
  ticks = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin()])}

\item{grid_col, axis_col}{grid & axis colors; both default to `#cccccc`}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\description{
modified from hrbrthemes
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ipsum()
}
}
\references{
\url{https://hrbrmstr.github.io/hrbrthemes/reference/theme_ipsum.html}
}
