% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAloc}
\alias{VAloc}
\title{VA locations}
\format{A simple feature collection with 1937 rows and 45 variables:
\describe{
  \item{StationNumber}{Site ID}
  \item{FacilityName}{Site Name}
  \item{PrimaryService}{Primary service of site}
  \item{PrimaryAttribute}{Location type}
  \item{Mobile}{Whether facility is a Mobile VA Clinic}
  \item{Address1}{Address component}
  \item{Address2}{Address component}
  \item{Address3}{Address component}
  \item{PhoneNumber}{Site's phone number}
  \item{City}{City}
  \item{StateAbbreviation }{State abbreviation}
  \item{Zip}{ZIP Code}
  \item{CemeteryType}{Whether National or State}
  \item{CemeteryOperations}{Operations}
  \item{City}{City}
  \item{VAAdministration}{Administration ID}
  \item{Latitude}{Latitude}
  \item{Longitude}{Longitude}
  ...
  \item{geometry}{Geometry of shape}
}}
\source{
\url{https://www.va.gov/vetdata/maps.asp}
}
\usage{
VAloc
}
\description{
A dataset of geometry and attributes of Veteran Affairs locations as of 9/30/2016
Includes Veterans Health Administration:  Medical Centers (VAMC),
Community - Based Outpatient Clinics (CBOC), Community Living Centers (CLC),
Domiciliary Residential Rehabilitation Treatment Program (DRRTP), Vet Centers (VCTR).
Veterans Benefits Administration: Regional Offices (RO).
National Cemetery Administration:  National and State Cemeteries.
}
\keyword{datasets}
