% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visn}
\alias{visn}
\title{VHA geography - visn level}
\format{A simple feature collection with 18 rows and 4 variables:
\describe{
  \item{VISN}{Veteran Integrated Service Network ID}
  \item{Shape_Leng}{Length of shape}
  \item{Shape_Area}{Area of shape}
  \item{geometry}{Geometry of shape}
}}
\source{
\url{https://www.data.va.gov/dataset/veterans-integrated-services-networks-visn-markets-submarkets-sectors-and-counties}
}
\usage{
visn
}
\description{
A dataset geometry and attributes of VHA geographies at the visn level
}
\keyword{datasets}
