% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{submarket}
\alias{submarket}
\title{VHA geography - submarket level}
\format{A simple feature collection with 153 rows and 6 variables:
\describe{
  \item{VISN}{Veteran Integrated Service Network ID}
  \item{MARKET}{VHA Market ID}
  \item{SUBMARKET}{VHA Submarket ID}
  \item{Shape_Leng}{Length of shape}
  \item{Shape_Area}{Area of shape}
  \item{geometry}{Geometry of shape}
}}
\source{
\url{https://www.data.va.gov/dataset/veterans-integrated-services-networks-visn-markets-submarkets-sectors-and-counties}
}
\usage{
submarket
}
\description{
A dataset of geometry and attributes of VHA geographies at the submarket level
}
\keyword{datasets}
