% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAloc}
\alias{VAloc}
\title{VA locations}
\format{A simple feature collection with 1953 rows and 50 variables:
\describe{
  \item{StationNumber}{Site ID}
  \item{FacilityName}{Site Name}
  \item{PrimaryService}{Primary service of site}
  \item{PrimaryAttribute}{Location type}
  \item{Mobile}{Whether facility is/has a Mobile VA Clinic}
  \item{Address1}{Address component}
  \item{Address2}{Address component}
  \item{Address3}{Address component}
  \item{PhoneNumber}{Site's phone number}
  \item{FaxNumber}{Site's fax number}
  \item{City}{City}
  \item{StateAbbreviation}{State abbreviation}
  \item{Zip}{ZIP Code}
  \item{CemeteryType}{Whether National or State}
  \item{CemeteryOperations}{Operations}
  \item{VAAdministration}{VA Administration Agency}
  \item{Latitude}{Latitude}
  \item{Longitude}{Longitude}
  \item{SuppService}{Supplemental Services}
  ...

  \item{STCD}{Combined ID for State FIPS and Congressional District number}
  \item{StateFIPS}{State FIPS Code}
  \item{CD}{Congressional District where facility is located}
  \item{LASTNAME}{U.S. Representative - Last Name}
  \item{FIRSTNAME}{U.S. Representative - First Name}
  \item{FIPS}{County FIPS Code}
  \item{CountyName}{County Name where facility is located}
  \item{VISN}{Veteran Integrated Service Network ID}
  \item{MARKET}{VHA Market ID}
  \item{SUBMARKET}{VHA Submarket ID}
  \item{SECTOR}{VHA Sector ID}
  \item{PrimaryRUCACode2010}{Primary Rural-Urban Commuting Area Code (2010 Census) of site's location}
  \item{SecondaryRUCACode2010}{Secondary Rural-Urban Commuting Area Code (2010 Census) of site's location*}
  \item{RUCA_UrbanRural}{Commuting Area Rural/Urban Classification*}
  \item{VADistrictCode}{VA Regional Districts: 1-Northeast, 2-Southeast, 3-Midwest, 4-Continental, 5-Pacific**}
  \item{geometry}{Geometry of shape}
}}
\source{
\url{https://www.va.gov/vetdata/maps.asp}
}
\usage{
VAloc
}
\description{
A dataset of geometry and attributes of Veteran Affairs locations as of 9/30/2018.
Veterans Health Administration (VHA):
Medical Centers (VAMC), Community - Based Outpatient Clinics (CBOC),
Community Living Centers (CLC), Domiciliary Residential Rehabilitation Treatment Program (DRRTP),
Vet Centers (VCTR).
}
\details{
Veterans Benefits Administration (VBA):  Regional Offices (RO).

National Cemetery Administration (NCA):  National and State Cemeteries.
}
\section{Notes}{

*2010 Secondary RUCA codes were revised on 7/3/2019 affecting 10,909 census tracts. \url{https://www.ers.usda.gov/data-products/rural-urban-commuting-area-codes/}

**In early 2019 VA Districts were restructured from five to four. \url{https://www.va.gov/opa/pressrel/pressrelease.cfm?id=5194}
}

\keyword{datasets}
